requirements.txt
viberbot==1.0.11
msgpack==0.5.6
pyyaml==6.0
gunicorn==20.0.4
flask_cors==3.0.10
flask==2.0.2
redis==3.4.1
